+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Unreleased
 * Save image with Exif metadata
 * Convert bytes -> human readable sizes
+* Check corrupted images
+* Convert items to list/tuple
+* Init everything via `from upyog.all import *`
+
 
 ## 0.5.5 -- 1 September 2022
 * Add a bunch of utilities related to iterating in `utils.utils`